refactor(AuthController): replace legacy controllers with Annotation based Controller
